+++ index.html
@@ -14,10 +14,11 @@
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
   <link href="https://fonts.googleapis.com/css2?family=Caveat&family=Open+Sans&display=swap" rel="stylesheet" />
   <link rel="stylesheet" href="css/styles.css" />
-  <script src="script.js" defer></script>
+  <script type="module" src="greetings_modules.js"></script>
+  <script type="module" src="script.js"></script>
 </head>
 
-<body class="home" onload="greetings()">
+<body class="home">
   <!-- //  INTRO  // -->
   <div class="content-wrapper">
     <div class="content-bg">
