+++ script.js
@@ -1,32 +1,6 @@
-// Greetings(
-//   (function () {
-//     let text = "";
-//     const d = new Date();
-//     const time = d.getHours();
-//     if (time < 10) {
-//       text = `Good Morning!`;
-//     } else if (time < 18) {
-//       text = `Good Day!`;
-//     } else {
-//       text = `Good Evening!`;
-//     }
-//     document.getElementById("greeting").innerHTML = text;
-//   })()
-// );
+import { greetings } from "./greetings_modules.js";
 
-const greetings = () => {
-  let text = "";
-  const d = new Date();
-  const time = d.getHours();
-  if (time < 10) {
-    text = `Good Morning!`;
-  } else if (time < 18) {
-    text = `Good Day!`;
-  } else {
-    text = `Good Evening!`;
-  }
-  document.getElementById("greeting").innerHTML = text;
-};
+document.getElementById("greeting").innerHTML = greetings();
 
 // Opens new browser windows
 let myGitWindow;
